+++ nodejs/fastWorker.js
@@ -75,6 +75,7 @@ const configKey = 'config';
 // Known good hashes for template sets
 const supportedHashes = {
     'bigip-fast-templates': [
+        '2d88c05f2b7ce83e595c42c780d51b1216c0cafcc027762f6f01990d2d43105a', // v1.21
         '8b5152db4930aa1f18f6c25b7c8a508c2901b35307333505c971de3ad5e26ef4', // v1.20
         '9b65dc0d08c673806e5d38ba2ce36215f71b0c25efe82708cdb0009e33559e6a', // v1.19
         '1f887241f2a27a4525cac4ed2c642a4cc020e655febc42f69913e6df1d6ab240', // v1.18